# [RFC]msg: batch handle sendmsg.