@majianpeng 2019-01-30
using #24678, limit iops=100K. 
using perf stat -p $fio sleep 10

- w/  this PR
       8131.414254      task-clock (msec)         #    0.813 CPUs utilized          
           153,754      context-switches          #    0.019 M/sec                  
                22      cpu-migrations            #    0.003 K/sec                  
               311      page-faults               #    0.038 K/sec                  
    22,583,193,034      cycles                    #    2.777 GHz                    
    29,743,137,521      instructions              #    1.32  insn per cycle         
     6,016,773,770      branches                  #  739.942 M/sec                  
        61,560,545      branch-misses             #    1.02% of all branches        

      10.001800038 seconds time elapsed

- w/o  this PR
       9231.558032      task-clock (msec)         #    0.923 CPUs utilized          
           152,000      context-switches          #    0.016 M/sec                  
                 6      cpu-migrations            #    0.001 K/sec                  
               179      page-faults               #    0.019 K/sec                  
    26,155,080,135      cycles                    #    2.833 GHz                    
    33,525,563,582      instructions              #    1.28  insn per cycle         
     6,809,907,324      branches                  #  737.677 M/sec                  
        75,640,166      branch-misses             #    1.11% of all branches

 CPUs utilized  can reduce from 0.923 to 0.813.