@rouming 2019-01-31
Instead of copying everything to a temporary buffer you could simply fill in iovec with the whole queue. That would be much more efficient (no copies, no lock/unlock sequences for each message, etc), take a look on one of my perf experiments patches here: https://github.com/rouming/ceph/commit/90831f241bccbbdc8ee1c26763c526b3773eed78

    before: IOPS=132k, BW=514MiB/s, Lat=0.973ms
       after: IOPS=166k, BW=650MiB/s, Lat=0.769ms

That is a good step forward.