@majianpeng 2019-02-26
@rouming . although your patch can improve performance better than my PR. But except PosixConnectedSocketImpl, there are NativeConnectedSocketImpl and RDMAConnectedSocketImpl. So we need add a abstract api like send function. But from the implementaion of  NativeConnectedSocketImpl and RDMAConnectedSocketImpl, i think it't hard to implement.