@rouming 2019-02-26
@majianpeng 
>But from the implementaion of NativeConnectedSocketImpl and RDMAConnectedSocketImpl, i think it't hard to implement.

My patch only shows the way, this is not a ready-to-commit solution. You always can pass messages queue to low-level and low-level implementation decides on its own how to deal with buffers: map to iovec or continue using them as bufferslist. 

BTW, did you check block sizes > 4k ? That what I've got on 128k:

 master:  IOPS=12.2k, BW=1524MiB/s, Lat=10.493
your PR: IOPS=9965, BW=1246MiB/s, Lat=12.841

Moreover, on small block sizes there is no any gain at all, in my case IOPS=108k for master and your PR.

Also you are saying, that you have a "limit iops=100K." - I bet you are testing protocolv2 only (by default), which indeed has some performance issue (@rjfd did you find any, btw?).