@majianpeng 2019-02-27
@rouming. I didn't do > 4k test.  I'll do that.  Thanks!