@majianpeng 2019-03-05
@rouming . sorry later to response. I rebase libfio_ceph_messenger to d7a279f7599e6917977f4770432fe3da7cfc47ed . And based on this do test, for bs=128K, write. master IOPS=8806, w/ my PR: IOPS=8807.