@rouming 2019-03-05
@majianpeng  Could you please test more than 1 bs and provide table like here, in the following commit description: https://github.com/rouming/ceph/commit/650c2c76da85a14dd38d2867cea5eaf88ffe7b0b

Also, since you are modifying two protocols would be nice to see how your change affects both of them.  Otherwise it is not clear what overall gain this patch brings.  Thanks.