@rouming 2019-03-05
@majianpeng I suggest you to play with crc, this is where I see the degradation on big blocks:

```
master, crc=on                            majianpeng
v1
 128k    IOPS=32.3k, BW=4033MiB/s     IOPS=22.8k, BW=2853MiB/s

v2
 128k   IOPS=31.4k, BW=3931MiB/s      IOPS=21.2k, BW=2649MiB/s
```