@majianpeng 2019-03-08
@rouming.  enable crd both for header and data; IODEPTH=128

- v1   | master   | w/ PR
--       |  ----- --  | --
4k      | 123000   | 112000
8k      | 97800     | 82700
32k    | 46200     | 40900
64k    | 24000     | 24400
128K  | 9030       | 13200
256k  | 2055       | 7190
1024k | 2055      | 2244

 For  larger block size, this PR show better performance

- v2     | master | w/ PR
--        | --         | --
4k       | 91200 | 92400
8k       | 81800 | 74400
32k     | 41000 | 36200
64k     | 32500 | 23800
128K   | 14400 | 13800
256k   | 7284   | 8131
1024   | 2653   | 2267

For  most case, master is better than this PR. 
Because v1 will drop. So from the result of v2,  this PR can decrease performance for some case. So i'll close this.