@markhpc 2019-03-14
The results look pretty variable to me.  I don't think we can draw a conclusion one way or the other about this PR based on the performance data.