@rouming 2019-03-14
@markhpc The problem with the current patch is that you can't infinitely bloat bufferlist on a deep queue, you have to limit iovecs and the size, otherwise that badly impacts the performance. Recently I submitted the following pr https://github.com/ceph/ceph/pull/26932, which avoids usage of bufferlist for posix stack and maps the whole message queue directly to iovec. That shows nice performance.